present를 controller로써 취급하도록 수정
