+++ Assets/Scenes/SampleScene.unity
@@ -616,7 +616,7 @@ GameObject:
   - component: {fileID: 613062957}
   - component: {fileID: 613062958}
   m_Layer: 0
-  m_Name: RoomTopPresent
+  m_Name: FurnitureScrollViewController
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -645,14 +645,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 613062956}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f9f7c42c2c33e3b44915ad4feb292e9e, type: 3}
+  m_Script: {fileID: 11500000, guid: 772d27f2a2e419146a773188fc096411, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  uiFurnitureInstallView: {fileID: 1448192702}
   uiFurnitureScrollViewView: {fileID: 204031311}
-  gridTilemapView: {fileID: 1741522679}
-  installFurnitureModel: {fileID: 0}
-  furnitureManagerModel: {fileID: 0}
 --- !u!1 &683175898
 GameObject:
   m_ObjectHideFlags: 0
@@ -684,6 +680,51 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 6
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1 &878539640
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 878539641}
+  - component: {fileID: 878539642}
+  m_Layer: 0
+  m_Name: GridInstallController
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &878539641
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 878539640}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 2136090767}
+  m_RootOrder: 4
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &878539642
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 878539640}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 49fa3fcb2c77b734aa7f93d554e68587, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  gridTilemapView: {fileID: 1741522679}
+  uiFurnitureInstallView: {fileID: 1448192702}
 --- !u!1 &907239253
 GameObject:
   m_ObjectHideFlags: 0
@@ -1041,7 +1082,7 @@ GameObject:
   - component: {fileID: 1214838968}
   - component: {fileID: 1214838969}
   m_Layer: 0
-  m_Name: FurniturePresent
+  m_Name: FurnitureController
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -1074,7 +1115,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   uiFurnitureStatusView: {fileID: 1942011225}
-  furnitureManagerModel: {fileID: 0}
 --- !u!1839735485 &1247916001 stripped
 Tilemap:
   m_CorrespondingSourceObject: {fileID: 2385568380500919003, guid: 073f93c52eb75ef4c9d2955b055d4248, type: 3}
@@ -1207,7 +1247,7 @@ GameObject:
   - component: {fileID: 1483704948}
   - component: {fileID: 1483704949}
   m_Layer: 0
-  m_Name: TimeInfoPresent
+  m_Name: TimeInfoController
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -1236,11 +1276,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 1483704947}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 59bd2a80c0ce82a438bf37403a333f7a, type: 3}
+  m_Script: {fileID: 11500000, guid: f97300d579f608345a4025502dff609f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   uiTimeView: {fileID: 1761500643}
-  timeInfoModel: {fileID: 0}
 --- !u!1 &1500267762
 GameObject:
   m_ObjectHideFlags: 0
@@ -1749,7 +1788,7 @@ GameObject:
   - component: {fileID: 2058236306}
   - component: {fileID: 2058236307}
   m_Layer: 0
-  m_Name: CharacterPresent
+  m_Name: CharacterController
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -1782,7 +1821,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   characterAIStatusUIView: {fileID: 907239255}
-  timeInfoModel: {fileID: 0}
   characterGameObject: {fileID: 683175898}
 --- !u!1 &2100911424
 GameObject:
@@ -1959,7 +1997,7 @@ GameObject:
   m_Component:
   - component: {fileID: 2136090767}
   m_Layer: 0
-  m_Name: Presents
+  m_Name: Controllers
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -1980,6 +2018,7 @@ Transform:
   - {fileID: 1483704948}
   - {fileID: 2058236306}
   - {fileID: 1214838968}
+  - {fileID: 878539641}
   m_Father: {fileID: 0}
   m_RootOrder: 5
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
