--- Assets/Scripts/Model/Grid.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f571fcca94dafa043af4bd8412564eca
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
