--- Assets/Scripts/Present.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1df2ca43d81915040b6e05beb7b941f0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
