--- Assets/Scripts/Present/RoomTopPresent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f9f7c42c2c33e3b44915ad4feb292e9e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
